+++ src/main/java/bowling/domain/frame/Frame.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.frame;
 
 public class Frame {
