+++ src/main/java/bowling/domain/frame/exception/PitchResultCreateException.java
@@ -1,4 +1,4 @@
-package bowling.domain.exception;
+package bowling.domain.frame.exception;
 
 public class PitchResultCreateException extends RuntimeException {
