+++ src/test/java/bowling/domain/frame/PitchResultTest.java
@@ -1,9 +1,9 @@
-package bowling.domain;
+package bowling.domain.frame;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import bowling.domain.exception.PitchResultCreateException;
+import bowling.domain.frame.exception.PitchResultCreateException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
